go fields: introduce fields as a concept
The fields package abstracts away dealing with and coercing types for
struct fields.

A descriptor is responsible for keeping track of all type information
associated with a struct field.

It automatically dereferences the value and remembers that we're dealing
with pointers in order to reduce the permutations of value assignment in
dealing with the reflected values.

Now rather than having to check both values if they are pointers when
assigning (2*2 permutations, 4 code branches) you can simply know that
this side is always dereferenced and only assign pointers when